requirements.txt
# Core
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1

# LLM
openai==1.58.1

# SQL
sqlalchemy==2.0.36

# Database drivers
sqlalchemy-bigquery==1.12.0
google-cloud-bigquery==3.27.0
pymysql==1.1.1
psycopg2-binary==2.9.10
